requirements.txt
# ============================================================
#  CRYSTAL AI - Dependencies  v2.1
# ============================================================

# Core AI
ollama==0.6.1

# Config
pyyaml==6.0.2

# Voice — STT
faster-whisper==1.0.3
pyaudio==0.2.14

# Voice — TTS
sounddevice==0.4.6
numpy==1.26.4

# Wake word
openwakeword==0.6.0

# Web UI
flask
flask-socketio

# System monitoring
psutil

# ── Phase 3: PC Control ──────────────────────────────────────
pyautogui          # screenshots, keyboard, mouse control
pyperclip          # clipboard read/write
pycaw              # Windows audio volume control (Windows only)
comtypes           # required by pycaw